moved persistence REST resource into io.rest.core bundle
