+++ bundles/core/org.eclipse.smarthome.core.persistence/build.properties
@@ -1,6 +1,5 @@
 output.. = target/classes/
 bin.includes = META-INF/,\
-               OSGI-INF/,\
                .,\
                about.html
 source.. = src/main/java/,\
