+++ bundles/io/org.eclipse.smarthome.io.rest.core/META-INF/MANIFEST.MF
@@ -28,6 +28,7 @@ Import-Package: com.google.common.base,
  org.eclipse.smarthome.core.items.events,
  org.eclipse.smarthome.core.library.items,
  org.eclipse.smarthome.core.library.types,
+ org.eclipse.smarthome.core.persistence,
  org.eclipse.smarthome.core.thing,
  org.eclipse.smarthome.core.thing.dto,
  org.eclipse.smarthome.core.thing.link,
